Removed AOP aspects. Reorganizing packets, files, classes, etc. Reworking API for better error handling and better responses. Preparing for implementing OpenAPI 3 documentation. Removed user account controller and tests and related logic.

+++ src/main/java/org/pizzeria/fabulosa/web/controllers/user/UserAddressController.java
@@ -3,21 +3,20 @@ package org.pizzeria.fabulosa.web.controllers.user;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.pizzeria.fabulosa.entity.address.Address;
-import org.pizzeria.fabulosa.entity.error.Error;
-import org.pizzeria.fabulosa.services.user.UserAddressService;
-import org.pizzeria.fabulosa.web.aop.annotations.ValidateUserId;
-import org.pizzeria.fabulosa.web.constants.ApiResponses;
-import org.pizzeria.fabulosa.web.constants.ApiRoutes;
+import org.pizzeria.fabulosa.common.entity.address.Address;
+import org.pizzeria.fabulosa.security.utils.UserSecurity;
 import org.pizzeria.fabulosa.web.dto.api.Response;
-import org.pizzeria.fabulosa.web.dto.api.Status;
+import org.pizzeria.fabulosa.web.service.user.UserAddressService;
+import org.pizzeria.fabulosa.web.util.constant.ApiResponses;
+import org.pizzeria.fabulosa.web.util.constant.ApiRoutes;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Set;
-import java.util.UUID;
+
+import static org.pizzeria.fabulosa.web.util.ResponseUtils.error;
 
 @RequiredArgsConstructor
 @RestController
@@ -27,64 +26,48 @@ public class UserAddressController {
 
 	private final UserAddressService userAddressService;
 
-	@ValidateUserId
 	@GetMapping
 	public ResponseEntity<Response> findUserAddressListById(@PathVariable Long userId, HttpServletRequest request) {
-		Set<Address> userAddressList = userAddressService.findUserAddressListById(userId);
-
-		Response response = Response.builder()
-				.status(Status.builder()
-						.description(!userAddressList.isEmpty() ? HttpStatus.OK.name() : HttpStatus.NO_CONTENT.name())
-						.code(!userAddressList.isEmpty() ? HttpStatus.OK.value() : HttpStatus.NO_CONTENT.value())
-						.isError(false)
-						.build())
-				.payload(userAddressList.isEmpty() ? null : userAddressList)
-				.build();
-
-		return ResponseEntity.ok(response);
-	}
-
-	@ValidateUserId
-	@PostMapping
-	public ResponseEntity<Response> createUserAddress(
-			@RequestBody @Valid Address address, @PathVariable Long userId, HttpServletRequest request) {
-		boolean ok = userAddressService.addUserAddress(userId, address);
-
-		Response response = Response.builder()
-				.status(Status.builder()
-						.description(ok ? HttpStatus.CREATED.name() : HttpStatus.BAD_REQUEST.name())
-						.code(ok ? HttpStatus.CREATED.value() : HttpStatus.BAD_REQUEST.value())
-						.isError(!ok)
-						.build())
-				.build();
-
-		if (!ok) {
-			response.setError(Error.builder()
-					.id(UUID.randomUUID().getMostSignificantBits())
-					.cause(ApiResponses.ADDRESS_MAX_SIZE)
-					.origin(UserController.class.getSimpleName() + ".createUserAddress")
-					.path(request.getPathInfo())
-					.logged(false)
-					.fatal(false)
-					.build());
+		if (!UserSecurity.valid(userId)) {
+			return UserSecurity.deny(request);
 		}
 
-		return ResponseEntity.ok(response);
+		Set<Address> userAddressList = userAddressService.findUserAddressListById(userId);
+
+		if (userAddressList.isEmpty()) {
+			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+		}
+
+		return ResponseEntity.ok(Response.builder().payload(userAddressList).build());
 	}
 
-	@ValidateUserId
+	@PostMapping
+	public ResponseEntity<Response> createUserAddress(@RequestBody @Valid Address address, @PathVariable Long userId, HttpServletRequest request) {
+		if (!UserSecurity.valid(userId)) {
+			return UserSecurity.deny(request);
+		}
+
+		boolean ok = userAddressService.addUserAddress(userId, address);
+
+		if (!ok) {
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(error(this.getClass().getSimpleName(), ApiResponses.ADDRESS_MAX_SIZE, request.getPathInfo()));
+		}
+
+		return ResponseEntity.status(HttpStatus.CREATED).build();
+	}
+
 	@DeleteMapping(ApiRoutes.USER_ADDRESS_ID)
 	public ResponseEntity<Response> deleteUserAddress(@PathVariable Long addressId, @PathVariable Long userId, HttpServletRequest request) {
+		if (!UserSecurity.valid(userId)) {
+			return UserSecurity.deny(request);
+		}
+
 		boolean result = userAddressService.removeUserAddress(userId, addressId);
 
-		Response response = Response.builder()
-				.status(Status.builder()
-						.description(result ? HttpStatus.OK.name() : HttpStatus.NO_CONTENT.name())
-						.code(result ? HttpStatus.OK.value() : HttpStatus.NO_CONTENT.value())
-						.isError(false)
-						.build())
-				.build();
+		if (!result) {
+			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+		}
 
-		return ResponseEntity.ok(response);
+		return ResponseEntity.ok().build();
 	}
 }
